+++ eth_defi/event_reader/fast_json_rpc.py
@@ -48,5 +48,12 @@ def patch_web3(web3: Web3):
 
     This greatly improves JSON-RPC API access speeds, when fetching
     multiple and large responses.
+
+    Example:
+
+    .. code-block:: python
+
+        from eth_defi.event_reader.fast_json_rpc import patch_web3
+        patch_web3(web3)
     """
     patch_provider(web3.provider)